+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to Semantic Versioning (https://semver.org/spec/v2.0.0.
 
 - `config()` now supports `checks` parameter for validation functions, matching `io()`.
 
+### Changed
+
+- `config()` parameter `convert` is now deprecated and emits a warning.
+
 ## [0.3.51] - 2026-03-04
 
 ### Changed